Move plugin into main package
