+++ AxonDeepSeg/ads_napari/Settings_menu_ui.ui
