+++ AxonDeepSeg/ads_napari/__init__.py
@@ -1,5 +1,3 @@
-__version__ = "1.0.0"
-
 from ._widget import ADSplugin
 
 __all__ = (
