+++ AxonDeepSeg/ads_napari/settings_menu_ui.py
