--- plugins/napari_ADS/napari.yaml
@@ -1,9 +1,0 @@
-name: napari-ADS
-contributions:
-  commands:
-    - id: napari-ADS.make_qwidget
-      python_name: napari_ADS._widget:ADSplugin
-      title: Make ADS plugin
-  widgets:
-    - command: napari-ADS.make_qwidget
-      display_name: ADS plugin
