--- plugins/setup.cfg
@@ -1,19 +1,0 @@
-[metadata]
-name = napari-ADS
-version = 1.0.0
-classifiers =
-    Framework :: napari
-
-[options]
-packages = find:
-include_package_data = True
-install_requires =
-    napari[all]
-    qtpy
-
-[options.entry_points]
-napari.manifest =
-    napari-ADS = napari_ADS:napari.yaml
-
-[options.package_data]
-napari-ADS = napari.yaml
